Move validation to module
